+++ src/Task/GeneratePhpstanConfig.php
@@ -1,0 +1,31 @@
+<?php
+
+/**
+ * @package Effigy
+ * @license http://opensource.org/licenses/MIT
+ */
+
+declare(strict_types=1);
+
+namespace DecodeLabs\Effigy\Task;
+
+use DecodeLabs\Atlas\File;
+use DecodeLabs\Clip\Task;
+use DecodeLabs\Effigy;
+use DecodeLabs\Effigy\Task\GeneratePhpstanConfig\PhpstanTemplate;
+use DecodeLabs\Effigy\Template;
+
+class GeneratePhpstanConfig implements Task
+{
+    use GenerateFileTrait;
+
+    protected function getTargetFile(): File
+    {
+        return Effigy::$rootDir->getFile('phpstan.neon');
+    }
+
+    protected function getTemplate(): Template
+    {
+        return new PhpstanTemplate();
+    }
+}